+++ src/backend/base/langflow/api/v2/workflow.py
@@ -1,42 +1,36 @@
 """V2 Workflow execution endpoints.
 
-This module implements the V2 Workflow API endpoints for executing flows with
-enhanced error handling, timeout protection, and structured responses.
+This module implements the V2 Workflow API route handlers for executing flows
+with enhanced error handling, timeout protection, and structured responses.
+The execution machinery itself lives in sibling modules:
+
+    - ``workflow_validation``: request/permission guards.
+    - ``workflow_execution``: sync + streaming run-driving.
+    - ``workflow_background``: durable, re-attachable background runs.
 
 Endpoints:
-    POST /workflow: Execute a workflow (sync, stream, or background modes)
-    GET /workflow: Get workflow job status by job_id
-    POST /workflow/stop: Stop a running workflow execution
+    POST /workflows: Execute a workflow (sync, stream, or background modes)
+    GET /workflows: Get workflow job status by job_id
+    POST /workflows/stop: Stop a running workflow execution
+    GET /workflows/{job_id}/events: Re-attach to a background run's event stream
 
 Features:
     - Comprehensive error handling with structured error responses
     - Timeout protection for long-running executions
     - Support for multiple execution modes (sync, stream, background)
     - Session-cookie or API-key authentication
-
-Configuration:
-    EXECUTION_TIMEOUT: Maximum execution time for synchronous workflows (300 seconds)
 """
 
 from __future__ import annotations
 
 import asyncio
 import contextlib
-import json
-import time
-from collections.abc import AsyncIterator, Callable, Iterable
-from copy import deepcopy
 from typing import Annotated
 from uuid import UUID, uuid4
 
-from ag_ui.core import CustomEvent
 from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query, Request, status
 from fastapi.responses import EventSourceResponse, StreamingResponse
-from fastapi.sse import format_sse_event
-from lfx.events.event_manager import create_default_event_manager
-from lfx.graph.graph.base import Graph
 from lfx.log.logger import logger
-from lfx.schema.schema import InputValueRequest
 from lfx.schema.workflow import (
     WORKFLOW_EXECUTION_RESPONSES,
     WORKFLOW_STATUS_RESPONSES,
@@ -49,28 +43,35 @@ from lfx.schema.workflow import (
     WorkflowStopResponse,
 )
 from lfx.services.deps import injectable_session_scope_readonly
-from lfx.utils.flow_validation import CustomComponentValidationError, validate_flow_for_current_settings
 from pydantic_core import ValidationError as PydanticValidationError
 from sqlalchemy.exc import OperationalError
 
-from langflow.api.utils import extract_global_variables_from_headers
-from langflow.api.v1.schemas import FlowDataRequest, RunResponse
 from langflow.api.v2.adapters import (
     STREAM_ADAPTERS,
-    StreamAdapter,
     StreamAdapterContext,
-    StreamEvent,
     UnknownStreamProtocolError,
     available_protocols,
     get_stream_adapter,
 )
-from langflow.api.v2.converters import (
-    ParsedWorkflowRun,
-    create_error_response,
-    parse_workflow_run_request,
-    run_response_to_workflow_response,
+from langflow.api.v2.converters import parse_workflow_run_request
+from langflow.api.v2.workflow_background import (
+    _BACKGROUND_RUNS,
+    _cancel_workflow_queue_job,
+    _finish_cancelled_background_run,
+    execute_workflow_background,
+)
+from langflow.api.v2.workflow_execution import (
+    _execute_streaming_workflow,
+    _resolve_execution_timeout,
+    execute_sync_workflow_with_timeout,
 )
 from langflow.api.v2.workflow_reconstruction import reconstruct_workflow_response_from_job_id
+from langflow.api.v2.workflow_validation import (
+    _enforce_flow_data_override_owner,
+    _flow_not_found_privacy_exception,
+    _reject_unsupported_sync_fields,
+    _validate_flow_data_for_execution,
+)
 from langflow.exceptions.api import (
     WorkflowQueueFullError,
     WorkflowResourceError,
@@ -79,38 +80,13 @@ from langflow.exceptions.api import (
     WorkflowValidationError,
 )
 from langflow.helpers.flow import get_flow_by_id_or_endpoint_name
-from langflow.processing.process import process_tweaks, run_graph_internal
 from langflow.services.auth.utils import get_current_user_for_workflow
 from langflow.services.authorization import FlowAction, ensure_flow_permission
-from langflow.services.authorization.fetch import deny_to_404
-from langflow.services.database.models.flow.model import FlowRead
 from langflow.services.database.models.jobs.model import JobType
 from langflow.services.database.models.user.model import UserRead
-from langflow.services.deps import get_job_service, get_memory_base_service, get_queue_service, get_task_service
-
-# Configuration constants
-EXECUTION_TIMEOUT = 300  # 5 minutes default timeout for sync execution
-
+from langflow.services.deps import get_job_service
 
 router = APIRouter(prefix="/workflows", tags=["Workflow"])
-
-
-async def generate_flow_events(*args, **kwargs) -> None:
-    """Lazily call the v1 build stream to avoid import cycles during router setup."""
-    from langflow.api.build import generate_flow_events as _generate_flow_events
-
-    await _generate_flow_events(*args, **kwargs)
-
-
-async def _cancel_workflow_queue_job(job_id: str) -> bool:
-    """Lazily call the shared build cancellation path to avoid import cycles."""
-    from langflow.api.build import cancel_flow_build
-    from langflow.services.job_queue.service import JobQueueNotFoundError
-
-    try:
-        return await cancel_flow_build(job_id=job_id, queue_service=get_queue_service())
-    except JobQueueNotFoundError:
-        return False
 
 
 def _unknown_protocol_http_exception(exc: UnknownStreamProtocolError) -> HTTPException:
@@ -130,115 +106,6 @@ def _unknown_protocol_http_exception(exc: UnknownStreamProtocolError) -> HTTPExc
             "available": exc.available,
         },
     )
-
-
-def _flow_not_found_privacy_exception(exc: HTTPException, flow_id: str) -> HTTPException:
-    return deny_to_404(exc, detail=f"Flow with id {flow_id} not found")
-
-
-def _reject_unsupported_sync_fields(parsed: ParsedWorkflowRun) -> None:
-    """Reject request fields the inline sync path does not execute."""
-    if parsed.mode != "sync":
-        return
-
-    unsupported_fields: list[str] = []
-    if parsed.data is not None:
-        unsupported_fields.append("data")
-    if parsed.files:
-        unsupported_fields.append("files")
-    if parsed.start_component_id is not None:
-        unsupported_fields.append("start_component_id")
-    if parsed.stop_component_id is not None:
-        unsupported_fields.append("stop_component_id")
-
-    if unsupported_fields:
-        fields = ", ".join(unsupported_fields)
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail={
-                "error": "Unsupported sync request fields",
-                "code": "SYNC_MODE_UNSUPPORTED_FIELDS",
-                "message": f"mode='sync' does not support request fields: {fields}. Use mode='stream' or "
-                "mode='background' for live-canvas overrides, files, or partial-run boundaries.",
-                "fields": unsupported_fields,
-            },
-        )
-
-
-def _enforce_flow_data_override_owner(parsed: ParsedWorkflowRun, flow: FlowRead, current_user: UserRead) -> None:
-    """Only the flow owner may execute caller-supplied graph data."""
-    if parsed.data is None or flow.user_id == current_user.id:
-        return
-
-    raise _flow_not_found_privacy_exception(
-        HTTPException(
-            status_code=status.HTTP_403_FORBIDDEN,
-            detail="Only the flow owner can override flow data during execution",
-        ),
-        parsed.flow_id,
-    )
-
-
-def _validate_flow_data_for_execution(parsed: ParsedWorkflowRun, flow: FlowRead) -> None:
-    """Apply the same server-side component policy gate used by v1/public runs."""
-    try:
-        if parsed.data is not None:
-            validate_flow_for_current_settings(parsed.data)
-        elif flow.data:
-            validate_flow_for_current_settings(flow.data)
-    except CustomComponentValidationError as exc:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc)) from exc
-    except RuntimeError as exc:
-        raise HTTPException(status_code=status.HTTP_503_SERVICE_UNAVAILABLE, detail=str(exc)) from exc
-
-
-def _validate_output_ids(output_ids: list[str] | None, terminal_node_ids: list[str]) -> None:
-    """Reject ``output_ids`` that aren't outputs of this flow, BEFORE it runs.
-
-    Checks against the terminal node ids known after graph build but before
-    execution, so a typo or wrong id costs no compute. ``available`` lists the
-    flow's real output ids so callers can self-correct. None/empty means "no
-    selection" and never raises.
-    """
-    if not output_ids:
-        return
-    known = set(terminal_node_ids)
-    unknown = [output_id for output_id in output_ids if output_id not in known]
-    if unknown:
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail={
-                "error": "Unknown output_ids",
-                "code": "UNKNOWN_OUTPUT_IDS",
-                "message": f"output_ids not produced by this flow: {unknown}.",
-                "available": terminal_node_ids,
-            },
-        )
-
-
-def _build_run_inputs(parsed: ParsedWorkflowRun) -> list[InputValueRequest] | None:
-    """Build the graph input list from the AG-UI chat message, if any.
-
-    The last user message becomes a single chat input; an empty message means
-    the flow runs with no chat input (parameters arrive via tweaks instead).
-    """
-    if not parsed.input_value:
-        return None
-    return [InputValueRequest(components=[], input_value=parsed.input_value, type="chat")]
-
-
-def _resolve_request_variables(body_globals: dict[str, str], http_request: Request | None) -> dict[str, str]:
-    """Merge request-level global variables for a v2 workflow execution.
-
-    v2 workflows take globals from the JSON request body (``globals``). The
-    ``X-LANGFLOW-GLOBAL-VAR-*`` headers remain a supported transport (the
-    OpenAI-compatible Responses API passes globals that way); body globals win
-    on conflict.
-    """
-    header_globals: dict[str, str] = {}
-    if http_request is not None:
-        header_globals = extract_global_variables_from_headers(http_request.headers)
-    return {**header_globals, **dict(body_globals or {})}
 
 
 @router.post(
@@ -373,25 +240,27 @@ async def execute_workflow(
             ) from e
         raise
     except OperationalError as e:
+        await logger.aexception("Database error fetching flow for workflow execution")
         raise HTTPException(
             status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
             detail={
                 "error": "Service unavailable, Please try again.",
                 "code": "DATABASE_ERROR",
-                "message": f"Failed to fetch flow: {e!s}",
+                "message": "Failed to fetch flow. Please try again.",
                 "flow_id": parsed.flow_id,
             },
         ) from e
     except WorkflowTimeoutError:
+        timeout_seconds = _resolve_execution_timeout()
         raise HTTPException(
             status_code=status.HTTP_408_REQUEST_TIMEOUT,
             detail={
                 "error": "Execution timeout",
                 "code": "EXECUTION_TIMEOUT",
-                "message": f"Workflow execution exceeded {EXECUTION_TIMEOUT} seconds",
+                "message": f"Workflow execution exceeded {timeout_seconds} seconds",
                 "job_id": str(job_id),
                 "flow_id": str(parsed.flow_id),
-                "timeout_seconds": EXECUTION_TIMEOUT,
+                "timeout_seconds": timeout_seconds,
             },
         ) from None
     except (PydanticValidationError, WorkflowValidationError) as e:
@@ -425,798 +294,16 @@ async def execute_workflow(
             },
         ) from err
     except Exception as err:
+        await logger.aexception("Unexpected error during workflow execution")
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail={
                 "error": "Internal server error",
                 "code": "INTERNAL_SERVER_ERROR",
-                "message": f"An unexpected error occurred: {err!s}",
+                "message": "An unexpected error occurred.",
                 "flow_id": parsed.flow_id,
             },
         ) from err
-
-
-async def execute_sync_workflow_with_timeout(
-    parsed: ParsedWorkflowRun,
-    flow: FlowRead,
-    job_id: UUID,
-    current_user: UserRead,
-    background_tasks: BackgroundTasks,
-    http_request: Request,
-) -> WorkflowExecutionResponse:
-    """Execute workflow with timeout protection.
-
-    Args:
-        parsed: The parsed AG-UI run parameters
-        flow: The flow to execute
-        job_id: Generated job ID for tracking
-        current_user: Authenticated user
-        background_tasks: FastAPI background tasks
-        http_request: The HTTP request object for extracting headers
-
-    Returns:
-        WorkflowExecutionResponse with complete results
-
-    Raises:
-        WorkflowTimeoutError: If execution exceeds timeout
-        WorkflowValidationError: If flow validation fails
-    """
-    try:
-        return await asyncio.wait_for(
-            execute_sync_workflow(
-                parsed=parsed,
-                flow=flow,
-                job_id=job_id,
-                current_user=current_user,
-                background_tasks=background_tasks,
-                http_request=http_request,
-            ),
-            timeout=EXECUTION_TIMEOUT,
-        )
-    except asyncio.TimeoutError as e:
-        raise WorkflowTimeoutError from e
-
-
-async def execute_sync_workflow(
-    parsed: ParsedWorkflowRun,
-    flow: FlowRead,
-    job_id: UUID,
-    current_user: UserRead,
-    background_tasks: BackgroundTasks,  # noqa: ARG001
-    http_request: Request,
-) -> WorkflowExecutionResponse:
-    """Execute workflow synchronously and return complete results.
-
-    This function implements a two-tier error handling strategy:
-        1. System-level errors (validation, graph build): Raised as exceptions
-        2. Component execution errors: Returned in response body with HTTP 200
-
-    This approach allows clients to receive partial results even when some
-    components fail, which is useful for debugging and incremental processing.
-
-    Execution Flow:
-        1. Apply tweaks and chat input from the parsed AG-UI request
-        2. Validate flow data exists
-        3. Extract context from HTTP headers
-        4. Build graph from flow data with tweaks applied
-        5. Identify terminal nodes for execution
-        6. Execute graph and collect results
-        7. Convert V1 RunResponse to V2 WorkflowExecutionResponse
-
-    Args:
-        parsed: The parsed AG-UI run parameters with tweaks and chat input
-        flow: The flow model from database
-        job_id: Generated job ID for tracking this execution
-        current_user: Authenticated user for permission checks
-        background_tasks: FastAPI background tasks (unused in sync mode)
-        http_request: The HTTP request object for extracting headers
-
-    Returns:
-        WorkflowExecutionResponse: Complete execution results with outputs and metadata
-
-    Raises:
-        WorkflowValidationError: If flow data is None or graph build fails
-    """
-    # Tweaks and chat input come straight from the parsed AG-UI request
-    tweaks = parsed.tweaks
-    session_id = parsed.session_id
-
-    # Validate flow data - this is a system error, not execution error
-    if flow.data is None:
-        msg = f"Flow {flow.id} has no data. The flow may be corrupted."
-        raise WorkflowValidationError(msg)
-
-    # Resolve request-level variables: body ``globals`` plus the legacy
-    # X-LANGFLOW-GLOBAL-VAR-* headers (still used by the Responses API).
-    # Body globals win on conflict.
-    request_variables = _resolve_request_variables(parsed.globals, http_request)
-
-    # Build context from request variables (similar to V1's _run_flow_internal)
-    context = {"request_variables": request_variables} if request_variables else None
-
-    # Build graph - system error if this fails
-    try:
-        flow_id_str = str(flow.id)
-        user_id = str(current_user.id)
-        # Use deepcopy to prevent mutation of the original flow.data
-        # process_tweaks modifies nested dictionaries in-place
-        graph_data = deepcopy(flow.data)
-        graph_data = process_tweaks(graph_data, tweaks, stream=False)
-        # Pass context to graph (similar to V1's simple_run_flow)
-        # This allows components to access request metadata via graph.context
-        graph = Graph.from_payload(
-            graph_data, flow_id=flow_id_str, user_id=user_id, flow_name=flow.name, context=context
-        )
-        # Set run_id for tracing/logging (similar to V1's simple_run_flow)
-        graph.set_run_id(job_id)
-    except Exception as e:
-        msg = f"Failed to build graph from flow data: {e!s}"
-        raise WorkflowValidationError(msg) from e
-
-    # Get terminal nodes - these are the outputs we want
-    terminal_node_ids = graph.get_terminal_nodes()
-
-    # Validate request-side output selection BEFORE executing: a bad id must cost
-    # no compute. Raised outside the component-error try/except below, so it
-    # surfaces as a real 422 rather than a 200-with-failed body.
-    _validate_output_ids(parsed.output_ids, terminal_node_ids)
-
-    # Execute graph - component errors are caught and returned in response body
-    job_service = get_job_service()
-    await job_service.create_job(job_id=job_id, flow_id=flow_id_str, user_id=current_user.id)
-    try:
-        task_result, execution_session_id = await job_service.execute_with_status(
-            job_id=job_id,
-            run_coro_func=run_graph_internal,
-            graph=graph,
-            flow_id=flow_id_str,
-            session_id=session_id,
-            inputs=_build_run_inputs(parsed),
-            outputs=terminal_node_ids,
-            stream=False,
-        )
-
-        # Fire memory-base auto-capture hook — non-blocking background effect.
-        try:
-            _run_id_uuid = UUID(graph.run_id) if graph.run_id else None  # type-cast only; same run_id set on graph
-            await get_task_service().fire_and_forget_task(
-                get_memory_base_service().on_flow_output,
-                flow_id=flow.id,
-                session_id=execution_session_id,
-                job_id=_run_id_uuid,
-            )
-        except (RuntimeError, ValueError, OSError):
-            await logger.awarning("Memory base hook scheduling failed for flow %s", flow.id, exc_info=True)
-
-        # Build RunResponse
-        run_response = RunResponse(outputs=task_result, session_id=execution_session_id)
-        # Convert to WorkflowExecutionResponse
-        return run_response_to_workflow_response(
-            run_response=run_response,
-            flow_id=parsed.flow_id,
-            job_id=str(job_id),
-            inputs=parsed.tweaks,
-            graph=graph,
-            effective_globals=request_variables,
-            selected_ids=parsed.output_ids,
-        )
-
-    except asyncio.CancelledError:
-        # Re-raise CancelledError to allow timeout mechanism to work properly
-        # This ensures asyncio.wait_for() can properly cancel and raise TimeoutError
-        raise
-    except asyncio.TimeoutError as e:
-        # Re-raise TimeoutError to allow timeout mechanism to work properly
-        # This ensures asyncio.wait_for() can properly cancel and raise TimeoutError
-        raise WorkflowTimeoutError from e
-    except Exception as exc:  # noqa: BLE001
-        # Component execution errors - return in response body with HTTP 200
-        # This allows partial results and detailed error information per component
-        return create_error_response(
-            flow_id=parsed.flow_id,
-            job_id=job_id,
-            inputs=parsed.tweaks,
-            error=exc,
-            effective_globals=request_variables,
-        )
-
-
-def _single_input_value_request(parsed: ParsedWorkflowRun) -> InputValueRequest | None:
-    """Build the single chat InputValueRequest the v1 build loop accepts.
-
-    The v1 build path (``generate_flow_events``) takes a single
-    ``InputValueRequest``; when it receives ``None`` it falls back to
-    ``InputValueRequest(session=str(flow_id))``, which would wipe out the
-    caller's session id. We always return one with the parsed session so
-    component messages stay scoped to the user's active session, even when
-    there is no chat input (e.g. the playground "Run Flow" button).
-    """
-    if not parsed.session_id and not parsed.input_value:
-        return None
-    return InputValueRequest(
-        components=[],
-        input_value=parsed.input_value or "",
-        type="chat",
-        session=parsed.session_id,
-    )
-
-
-_QueueItem = tuple[str | None, bytes | None, float]
-
-
-class _WorkflowEventQueue:
-    """Bounded EventManager handoff that fails explicitly instead of dropping events."""
-
-    def __init__(self, maxsize: int) -> None:
-        self._queue: asyncio.Queue[_QueueItem] = asyncio.Queue(maxsize=maxsize)
-        self._overflowed = False
-        self._loop = asyncio.get_running_loop()
-        self._overflow_task: asyncio.Task[None] | None = None
-
-    @property
-    def maxsize(self) -> int:
-        return self._queue.maxsize
-
-    async def get(self) -> _QueueItem:
-        return await self._queue.get()
-
-    async def put(self, item: _QueueItem) -> None:
-        if self._overflowed:
-            return
-        await self._queue.put(item)
-
-    def put_nowait(self, item: _QueueItem) -> None:
-        if self._overflowed:
-            return
-        try:
-            self._queue.put_nowait(item)
-        except asyncio.QueueFull:
-            self._overflowed = True
-            self._overflow_task = self._loop.create_task(self._emit_overflow_error())
-
-    async def _emit_overflow_error(self) -> None:
-        payload = {
-            "event": "error",
-            "data": {
-                "error": "Workflow event stream exceeded buffering capacity; client is consuming events too slowly."
-            },
-        }
-        await self._queue.put((f"error-{uuid4()}", json.dumps(payload).encode("utf-8"), time.time()))
-        await self._queue.put((None, None, time.time()))
-
-    async def aclose(self) -> None:
-        if self._overflow_task is not None and not self._overflow_task.done():
-            self._overflow_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await self._overflow_task
-
-
-async def _stream_event_frames(
-    *,
-    adapter: StreamAdapter,
-    flow_id: UUID,
-    flow_name: str | None,
-    background_tasks: BackgroundTasks,
-    parsed: ParsedWorkflowRun,
-    current_user: UserRead,
-    source_flow_id: UUID | None = None,
-    run_id: str | None = None,
-    track_job_status: bool = True,
-) -> AsyncIterator[tuple[bytes, str]]:
-    """Run a flow via the v1 build-vertex loop, dispatch its events through ``adapter``.
-
-    Yields ``(sse_frame_bytes, event_type_str)`` pairs. The consumer
-    (streaming endpoint, background buffer) frames are pre-formatted with a
-    monotonic ``id:`` for ``Last-Event-ID`` resume. The ``event_type_str`` is
-    the adapter's protocol-native type so the buffer task can finalize a
-    background job's status structurally (no substring matching).
-
-    A failure during the run becomes a terminal protocol event (e.g.
-    ``RUN_ERROR`` for AG-UI, ``error`` for langflow) routed through the
-    adapter; closing the consumer cancels the run.
-
-    When the adapter is AG-UI, side-channel ``CustomEvent`` frames carry
-    the raw Langflow payload alongside the AG-UI translation for the
-    playground's chat-view. A follow-up retires this once chat-view
-    consumes the AG-UI ``TEXT_MESSAGE_*`` lifecycle directly.
-    """
-    # EventManager uses put_nowait(), so a plain bounded asyncio.Queue would
-    # silently drop frames via QueueFull. This adapter keeps memory bounded and
-    # converts overflow into an explicit stream error + sentinel.
-    queue = _WorkflowEventQueue(maxsize=_EVENT_QUEUE_MAX_SIZE)
-    event_manager = create_default_event_manager(queue)
-    input_request = _single_input_value_request(parsed)
-    flow_data = FlowDataRequest(**parsed.data) if parsed.data else None
-
-    # Captured from drive()'s exception path so the consumer can yield a
-    # guaranteed adapter.error_events(...) fallback after the queue loop ends.
-    # Layered error handling, by design:
-    #   1. ``event_manager.on_error(...)`` is the cooperative path: the
-    #      translator turns it into the protocol's terminal-error event (e.g.
-    #      RUN_ERROR for AG-UI, ``error`` for langflow) so the buffer's
-    #      structural detector flips the job to FAILED.
-    #   2. ``adapter.error_events(exc)`` is the dispatcher's guaranteed
-    #      fallback: emitted from the consumer side when ``generate_flow_events``
-    #      raises before any cooperative terminal error reaches the queue.
-    #      Without this yield, an early failure would leave the stream with no
-    #      terminal error event and the buffer would mark the job COMPLETED.
-    #   3. The buffer task's ``terminal_error_type`` check fires on either
-    #      RUN_ERROR source, so a single drive() failure cannot result in a
-    #      job marked COMPLETED.
-    drive_error: BaseException | None = None
-
-    async def drive() -> None:
-        nonlocal drive_error
-        try:
-            await generate_flow_events(
-                flow_id=flow_id,
-                background_tasks=background_tasks,
-                event_manager=event_manager,
-                inputs=input_request,
-                data=flow_data,
-                files=parsed.files,
-                stop_component_id=parsed.stop_component_id,
-                start_component_id=parsed.start_component_id,
-                # Persist vertex builds (keyed by ``run_id``) only for job-tracked
-                # runs so a background job's status can be reconstructed later. Live
-                # streams pass no ``run_id`` and keep the no-persist behavior.
-                log_builds=run_id is not None,
-                current_user=current_user,
-                flow_name=flow_name,
-                source_flow_id=source_flow_id,
-                run_id=run_id,
-                track_job_status=track_job_status,
-            )
-        except asyncio.CancelledError:
-            raise
-        except Exception as exc:  # noqa: BLE001
-            drive_error = exc
-            with contextlib.suppress(Exception):
-                await event_manager.queue.put((None, None, time.time()))
-        # generate_flow_events emits on_end and the sentinel on success.
-
-    def _frame(stream_event: StreamEvent, seq: int) -> tuple[bytes, str]:
-        return (
-            format_sse_event(data_str=stream_event.data_json, id=str(seq)),
-            stream_event.type,
-        )
-
-    # The AG-UI playground's chat-view consumes the v1 message payload via a
-    # side-channel ``CustomEvent``; emitted only when the wire protocol is
-    # AG-UI. A follow-up retires this once chat-view consumes AG-UI primitives.
-    emit_side_channel = adapter.name == "agui"
-    side_channel_events = frozenset({"add_message", "token", "remove_message", "error", "end"})
-    terminal_error_type = getattr(adapter, "terminal_error_type", None)
-    terminal_error_seen = False
-
-    seq = 0
-    run_task = asyncio.create_task(drive())
-    try:
-        for event in adapter.initial_events():
-            yield _frame(event, seq)
-            seq += 1
-        while True:
-            _, value, _ = await queue.get()
-            if value is None:
-                break
-            payload = json.loads(value.decode("utf-8"))
-            event_type = payload.get("event", "")
-            event_data = payload.get("data") or {}
-            if emit_side_channel and event_type in side_channel_events:
-                yield _frame(
-                    StreamEvent(
-                        type="CUSTOM",
-                        data_json=CustomEvent(
-                            name="langflow.event",
-                            value={"event_type": event_type, "data": event_data},
-                        ).model_dump_json(by_alias=True, exclude_none=True),
-                    ),
-                    seq,
-                )
-                seq += 1
-            for event in adapter.translate(event_type, event_data):
-                if terminal_error_type is not None and event.type == terminal_error_type:
-                    terminal_error_seen = True
-                yield _frame(event, seq)
-                seq += 1
-        for event in adapter.final_events():
-            if terminal_error_type is not None and event.type == terminal_error_type:
-                terminal_error_seen = True
-            yield _frame(event, seq)
-            seq += 1
-        # Guaranteed-fallback layer (see drive_error block above). If drive()
-        # captured an exception and no cooperative terminal error reached the
-        # stream, emit the adapter's terminal error event(s) here.
-        if drive_error is not None and not terminal_error_seen:
-            for event in adapter.error_events(drive_error):
-                if terminal_error_type is not None and event.type == terminal_error_type:
-                    terminal_error_seen = True
-                yield _frame(event, seq)
-                seq += 1
-    finally:
-        if not run_task.done():
-            run_task.cancel()
-        with contextlib.suppress(asyncio.CancelledError):
-            await run_task
-        await queue.aclose()
-
-
-def _execute_streaming_workflow(
-    *,
-    adapter: StreamAdapter,
-    parsed: ParsedWorkflowRun,
-    flow: FlowRead,
-    current_user: UserRead,
-    background_tasks: BackgroundTasks,
-) -> EventSourceResponse:
-    """Run a workflow live and stream events via ``adapter`` over server-sent events.
-
-    The graph is built inside ``generate_flow_events`` (the v1 build-vertex
-    loop) so the same per-vertex events the canvas already knows flow through
-    the adapter. A failure during the run becomes a terminal protocol event
-    routed through the adapter rather than an HTTP error.
-    """
-
-    async def _frames_only() -> AsyncIterator[bytes]:
-        async for frame, _event_type in _stream_event_frames(
-            adapter=adapter,
-            flow_id=flow.id,
-            flow_name=flow.name,
-            background_tasks=background_tasks,
-            parsed=parsed,
-            current_user=current_user,
-        ):
-            yield frame
-
-    return EventSourceResponse(
-        _frames_only(),
-        headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"},
-    )
-
-
-_CancelEventsFactory = Callable[[str], Iterable[StreamEvent]]
-
-
-class _BackgroundRun:
-    """In-memory buffer of a background run's protocol-native SSE frames for re-attach.
-
-    The buffer lives in the process; restarts drop it. Multiple readers can
-    re-attach concurrently and tail until the run ends. The frames are
-    already serialized in the protocol the original POST requested via
-    ``stream_protocol``; re-attach replays them as-is. Mixing protocols
-    across a single run is not supported.
-
-    Per-run frame count is bounded by ``_MAX_FRAMES_PER_BACKGROUND_RUN`` so a
-    long verbose run (token-by-token streams, repeated tool calls) cannot
-    exhaust process memory while ``_MAX_BACKGROUND_RUNS`` only caps the
-    number of buffers. When the cap is reached the oldest frames are
-    evicted; re-attach with ``Last-Event-ID`` past that point will start
-    from the new buffer head (replay loss is preferred over OOM).
-    """
-
-    def __init__(self, user_id: str, stream_protocol: str = "agui") -> None:
-        self.user_id = user_id
-        self.stream_protocol = stream_protocol
-        self._cancel_events: _CancelEventsFactory | None = None
-        self.frames: list[bytes] = []
-        # Index of the first frame still in ``frames`` (monotonic across the
-        # life of the buffer). Once eviction starts, ``frames[i]`` corresponds
-        # to logical event id ``base_index + i``.
-        self.base_index = 0
-        self.done = False
-        self._cond = asyncio.Condition()
-
-    def set_cancel_events(self, cancel_events: _CancelEventsFactory) -> None:
-        self._cancel_events = cancel_events
-
-    @property
-    def has_cancel_events(self) -> bool:
-        return self._cancel_events is not None
-
-    def _append_locked(self, frame: bytes) -> None:
-        self.frames.append(frame)
-        overflow = len(self.frames) - _MAX_FRAMES_PER_BACKGROUND_RUN
-        if overflow > 0:
-            del self.frames[:overflow]
-            self.base_index += overflow
-
-    async def append(self, frame: bytes) -> None:
-        async with self._cond:
-            self._append_locked(frame)
-            self._cond.notify_all()
-
-    async def finish(self) -> None:
-        async with self._cond:
-            self.done = True
-            self._cond.notify_all()
-
-    async def finish_cancelled(self, reason: str) -> None:
-        """Append cancellation terminal events and finish replay exactly once."""
-        async with self._cond:
-            if self.done:
-                return
-            events: list[StreamEvent] = []
-            if self._cancel_events is not None:
-                with contextlib.suppress(Exception):
-                    events = list(self._cancel_events(reason))
-            for event in events:
-                seq = self.base_index + len(self.frames)
-                self._append_locked(format_sse_event(data_str=event.data_json, id=str(seq)))
-            self.done = True
-            self._cond.notify_all()
-
-    async def replay(self, start_index: int) -> AsyncIterator[bytes]:
-        """Yield buffered frames from ``start_index`` and tail until done.
-
-        ``start_index`` is in logical event-id space (matches what was emitted
-        on ``id:`` lines). If the caller's ``Last-Event-ID`` points before the
-        buffer's current head (frames evicted under memory pressure), we
-        replay from the head and the caller observes a gap.
-        """
-        idx = max(start_index, 0)
-        while True:
-            async with self._cond:
-                head = self.base_index
-                tail = head + len(self.frames)
-                idx = max(idx, head)
-                while idx >= tail and not self.done:
-                    await self._cond.wait()
-                    head = self.base_index
-                    tail = head + len(self.frames)
-                    idx = max(idx, head)
-                snapshot = self.frames[idx - head :]
-                finished = self.done
-            for frame in snapshot:
-                yield frame
-            idx += len(snapshot)
-            if finished and idx >= self.base_index + len(self.frames):
-                return
-
-
-# Process-local registry of background runs keyed by job_id, bounded by
-# ``_MAX_BACKGROUND_RUNS`` (oldest evicted first). Re-attach reads this.
-_MAX_BACKGROUND_RUNS = 100
-# Per-run frame ceiling. Caps memory for a single long/verbose run so a
-# token-streaming flow can't exhaust the process. 10k frames covers minutes
-# of dense token streams with room to spare; beyond that we evict oldest.
-_MAX_FRAMES_PER_BACKGROUND_RUN = 10_000
-# Inline stream queue between the build loop and the SSE consumer. Bounded
-# so a slow consumer applies backpressure to the build loop instead of
-# letting frames accumulate without bound when the network is slow.
-_EVENT_QUEUE_MAX_SIZE = 256
-_BACKGROUND_RUNS: dict[str, _BackgroundRun] = {}
-_WORKFLOW_CANCELLED_MESSAGE = "Workflow run cancelled."
-
-
-async def _finalize_job_status(job_uuid: UUID, terminal_status: JobStatus) -> None:
-    """Update job status to a terminal value, but never overwrite CANCELLED.
-
-    ``stop_workflow`` sets the job to CANCELLED. The buffer task runs in
-    parallel and reaches its ``finally`` block shortly after; if it
-    unconditionally wrote COMPLETED/FAILED it would race with the cancellation
-    and silently overwrite the user's stop intent. Re-read the row first and
-    skip the update if a cancellation already landed.
-    """
-    job_service = get_job_service()
-    try:
-        job = await job_service.get_job_by_job_id(job_id=job_uuid)
-    except Exception:  # noqa: BLE001
-        job = None
-    if job is not None and job.status == JobStatus.CANCELLED:
-        return
-    with contextlib.suppress(Exception):
-        await job_service.update_job_status(
-            job_uuid,
-            terminal_status,
-            finished_timestamp=True,
-        )
-
-
-async def _clear_background_run(job_id: str) -> None:
-    """Pop the background run registry entry and wake any re-attach waiters.
-
-    Used for true cleanup paths that should discard replay state, such as
-    failed scheduling or test teardown. Cancellation normally finishes through
-    the owning buffer task so it can append protocol-native terminal events
-    before replay is marked done.
-    """
-    bg_run = _BACKGROUND_RUNS.pop(job_id, None)
-    if bg_run is not None:
-        await bg_run.finish()
-
-
-async def _finish_cancelled_background_run(job_id: str) -> None:
-    """Finish a stopped local run when no owner task is available to do it."""
-    bg_run = _BACKGROUND_RUNS.get(job_id)
-    if bg_run is None or bg_run.done:
-        return
-    if not bg_run.has_cancel_events:
-        with contextlib.suppress(UnknownStreamProtocolError):
-            adapter = get_stream_adapter(
-                bg_run.stream_protocol,
-                StreamAdapterContext(run_id=job_id, thread_id=job_id),
-            )
-            bg_run.set_cancel_events(adapter.cancel_events)
-    await bg_run.finish_cancelled(_WORKFLOW_CANCELLED_MESSAGE)
-
-
-async def _register_background_run(job_id: str, bg_run: _BackgroundRun) -> None:
-    """Register a background run, evicting a completed entry when full.
-
-    Prefer evicting the oldest *completed* run so a long-running job's
-    re-attach handle survives. If every slot is still active, evict the
-    oldest one anyway to keep the registry bounded, cancel its still-running
-    buffer writer so it stops appending into a run no reader can find, and log
-    a warning.
-    """
-    if len(_BACKGROUND_RUNS) >= _MAX_BACKGROUND_RUNS:
-        evict_key = next(
-            (key for key, run in _BACKGROUND_RUNS.items() if run.done),
-            None,
-        )
-        evicted_running = False
-        if evict_key is None:
-            evict_key = next(iter(_BACKGROUND_RUNS))
-            evicted_running = True
-            logger.warning(
-                "Background run registry full with no completed entries; "
-                "evicting still-running job %s to make room for %s",
-                evict_key,
-                job_id,
-            )
-        _BACKGROUND_RUNS.pop(evict_key, None)
-        if evicted_running:
-            # Stop the orphaned buffer writer: without its registry entry no
-            # reader can find it, but the coroutine keeps appending frames.
-            # cancel_flow_build raises CancelledError into the build loop, which
-            # runs the buffer's finally/finish_cancelled path.
-            with contextlib.suppress(Exception):
-                await _cancel_workflow_queue_job(evict_key)
-    _BACKGROUND_RUNS[job_id] = bg_run
-
-
-async def _buffer_background_run(
-    *,
-    bg_run: _BackgroundRun,
-    flow: FlowRead,
-    parsed: ParsedWorkflowRun,
-    job_id: str,
-    current_user: UserRead,
-    stream_protocol: str,
-) -> None:
-    """Run a background flow, buffer its frames, and finalize job status.
-
-    The adapter is resolved from ``stream_protocol`` so re-attach replays
-    frames in the protocol the caller requested. Validation of
-    ``stream_protocol`` happens at the route handler; this function still
-    guards against ``UnknownStreamProtocolError`` because it runs as a
-    fire-and-forget background coroutine. If adapter registration breaks
-    between the route's check and this call (e.g. a registry mutation), we
-    flip the job to FAILED and exit cleanly rather than dying silently and
-    leaving the job stuck at QUEUED.
-
-    The buffer's terminal-status detection keys off
-    ``adapter.terminal_error_type``.
-    """
-    try:
-        adapter = get_stream_adapter(
-            stream_protocol,
-            StreamAdapterContext(
-                run_id=parsed.run_id or job_id,
-                thread_id=parsed.session_id or str(flow.id),
-            ),
-        )
-    except UnknownStreamProtocolError:
-        # Fire-and-forget coroutine: do not raise, the route already returned.
-        await bg_run.finish()
-        job_uuid = UUID(job_id) if isinstance(job_id, str) else job_id
-        await _finalize_job_status(job_uuid, JobStatus.FAILED)
-        return
-
-    bg_run.set_cancel_events(adapter.cancel_events)
-    terminal_error_type = adapter.terminal_error_type
-    fresh_background_tasks = BackgroundTasks()
-    errored = False
-    cancelled = False
-    job_uuid = UUID(job_id) if isinstance(job_id, str) else job_id
-    try:
-        with contextlib.suppress(Exception):
-            await get_job_service().update_job_status(job_uuid, JobStatus.IN_PROGRESS)
-        async for frame, event_type in _stream_event_frames(
-            adapter=adapter,
-            flow_id=flow.id,
-            flow_name=flow.name,
-            background_tasks=fresh_background_tasks,
-            parsed=parsed,
-            current_user=current_user,
-            # Build under the job id so the run's vertex builds are persisted
-            # keyed by job_id and GET-status reconstruction can find them.
-            run_id=job_id,
-            track_job_status=False,
-        ):
-            if terminal_error_type is not None and event_type == terminal_error_type:
-                errored = True
-            await bg_run.append(frame)
-    except asyncio.CancelledError:
-        cancelled = True
-        await bg_run.finish_cancelled(_WORKFLOW_CANCELLED_MESSAGE)
-        raise
-    finally:
-        if not cancelled:
-            await bg_run.finish()
-        # ``generate_flow_events`` queues telemetry, tracing teardown, and
-        # other callbacks on this ``BackgroundTasks`` instance. In FastAPI's
-        # request lifecycle those run after the response is sent; the
-        # background path has no response carrying them, so drain the queue
-        # explicitly. Suppressed so a single failing telemetry callback does
-        # not derail job-status finalization.
-        with contextlib.suppress(Exception):
-            await fresh_background_tasks()
-        if not cancelled:
-            # ``update_job_status`` queries the Job table by its UUID primary key;
-            # passing the raw string would silently miss every row.
-            await _finalize_job_status(job_uuid, JobStatus.FAILED if errored else JobStatus.COMPLETED)
-
-
-async def execute_workflow_background(
-    parsed: ParsedWorkflowRun,
-    flow: FlowRead,
-    job_id: JobId,
-    current_user: UserRead,
-    http_request: Request,  # noqa: ARG001
-    stream_protocol: str,
-) -> WorkflowJobResponse:
-    """Run a workflow in the background, buffering protocol-native events for re-attach.
-
-    A job row is created so ``GET /workflows`` and ``POST /workflows/stop`` keep
-    working. The buffer task is scheduled through the queue service under
-    ``job_id`` so ``/stop`` can revoke it. Graph construction happens inside
-    the v1 build-vertex loop driven by ``_stream_event_frames`` with the
-    adapter selected by ``stream_protocol``; re-attach replays the frames in
-    that same protocol's wire shape.
-    """
-    try:
-        flow_id_str = str(flow.id)
-        job_id_str = str(job_id)
-
-        await get_job_service().create_job(
-            job_id=job_id,
-            flow_id=flow_id_str,
-            user_id=current_user.id,
-        )
-
-        bg_run = _BackgroundRun(user_id=str(current_user.id), stream_protocol=stream_protocol)
-        await _register_background_run(job_id_str, bg_run)
-
-        try:
-            queue_service = get_queue_service()
-            queue_service.create_queue(job_id_str)
-            queue_service.start_job(
-                job_id_str,
-                _buffer_background_run(
-                    bg_run=bg_run,
-                    flow=flow,
-                    parsed=parsed,
-                    job_id=job_id_str,
-                    current_user=current_user,
-                    stream_protocol=stream_protocol,
-                ),
-            )
-        except BaseException:
-            # If queue creation or scheduling fails after the bg_run is
-            # registered, the buffer would stay live with ``done=False`` and
-            # any re-attach client would block on ``_cond.wait()`` forever
-            # (the task that would call ``finish()`` was never scheduled).
-            # Clear the registry, mark the job FAILED, then re-raise.
-            await _clear_background_run(job_id_str)
-            await _finalize_job_status(job_id, JobStatus.FAILED)
-            raise
-        return WorkflowJobResponse(job_id=job_id_str, flow_id=parsed.flow_id, status=JobStatus.QUEUED)
-
-    except (WorkflowResourceError, WorkflowServiceUnavailableError, WorkflowQueueFullError):
-        raise
-    except MemoryError as exc:
-        raise WorkflowResourceError from exc
 
 
 @router.get(
@@ -1264,12 +351,13 @@ async def get_workflow_status(
     try:
         job = await job_service.get_job_by_job_id(job_id=job_id, user_id=current_user.id)
     except Exception as exc:
+        await logger.aexception("Database error retrieving workflow job %s", job_id)
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail={
                 "error": "Internal server error",
                 "code": "INTERNAL_SERVER_ERROR",
-                "message": f"Failed to retrieve job from database: {exc!s}",
+                "message": "Failed to retrieve job. Please try again.",
             },
         ) from exc
 
@@ -1359,24 +447,26 @@ async def get_workflow_status(
     except HTTPException:
         raise
     except WorkflowTimeoutError as err:
+        timeout_seconds = _resolve_execution_timeout()
         raise HTTPException(
             status_code=status.HTTP_408_REQUEST_TIMEOUT,
             detail={
                 "error": "Execution timeout",
                 "code": "EXECUTION_TIMEOUT",
-                "message": f"Workflow execution exceeded {EXECUTION_TIMEOUT} seconds",
+                "message": f"Workflow execution exceeded {timeout_seconds} seconds",
                 "job_id": job_id_str,
                 "flow_id": flow_id_str,
-                "timeout_seconds": EXECUTION_TIMEOUT,
+                "timeout_seconds": timeout_seconds,
             },
         ) from err
     except Exception as exc:
+        await logger.aexception("Unexpected error processing workflow job status for %s", job_id_str)
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail={
                 "error": "Internal server error",
                 "code": "INTERNAL_SERVER_ERROR",
-                "message": f"Failed to process job status: {exc!s}",
+                "message": "Failed to process job status.",
             },
         ) from exc
 
@@ -1414,12 +504,13 @@ async def stop_workflow(
         # 1. Fetch Job
         job = await job_service.get_job_by_job_id(job_id, user_id=current_user.id)
     except Exception as exc:
+        await logger.aexception("Database error retrieving workflow job %s for stop", job_id)
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail={
                 "error": "Internal server error",
                 "code": "INTERNAL_SERVER_ERROR",
-                "message": f"Failed to retrieve job status: {exc!s}",
+                "message": "Failed to retrieve job status. Please try again.",
             },
         ) from exc
 
@@ -1486,12 +577,13 @@ async def stop_workflow(
     except HTTPException:
         raise
     except Exception as exc:
+        await logger.aexception("Unexpected error stopping workflow job %s", job_id)
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail={
                 "error": "Internal server error",
                 "code": "INTERNAL_SERVER_ERROR",
-                "message": f"Failed to stop job: {job_id} - {exc!s}",
+                "message": f"Failed to stop job {job_id}.",
             },
         ) from exc
 
@@ -1559,9 +651,8 @@ async def reattach_workflow_events(
                 "error": "Background event buffer unavailable",
                 "code": "BACKGROUND_EVENTS_UNAVAILABLE",
                 "message": (
-                    f"Buffered events for job {job_id} are not available on this worker. "
-                    "Use the workflow status endpoint for this job, or route event re-attach "
-                    "requests to the worker that accepted the background run."
+                    f"Buffered events for job {job_id} are not available here. "
+                    "Use the workflow status endpoint for this job."
                 ),
                 "job_id": job_id,
             },
