+++ src/backend/tests/unit/api/v2/test_workflow.py
@@ -21,6 +21,7 @@ from langflow.services.database.models.flow.model import Flow
 from langflow.services.database.models.jobs.model import Job, JobType
 from lfx.schema.workflow import JobStatus
 from lfx.services.deps import session_scope
+from sqlalchemy.exc import OperationalError
 
 
 class TestWorkflowDeveloperAPIProtection:
@@ -322,7 +323,7 @@ class TestWorkflowStop:
 
         with (
             patch("langflow.api.v2.workflow.get_job_service") as mock_get_job_service,
-            patch("langflow.api.v2.workflow.get_queue_service", return_value=MissingQueueService()),
+            patch("langflow.api.v2.workflow_background.get_queue_service", return_value=MissingQueueService()),
         ):
             mock_job_service = MagicMock()
             mock_job_service.get_job_by_job_id = AsyncMock(return_value=mock_job)
@@ -737,7 +738,7 @@ class TestWorkflowIDORProtection:
 
         try:
             # Mock only the task service to prevent real background execution
-            with patch("langflow.api.v2.workflow.get_task_service") as mock_task_svc:
+            with patch("langflow.api.v2.workflow_execution.get_task_service") as mock_task_svc:
                 mock_task_service = MagicMock()
                 mock_task_service.fire_and_forget_task = AsyncMock()
                 mock_task_svc.return_value = mock_task_service
@@ -792,20 +793,20 @@ class TestValidateOutputIds:
     """
 
     def test_no_selection_is_a_noop(self):
-        from langflow.api.v2.workflow import _validate_output_ids
+        from langflow.api.v2.workflow_validation import _validate_output_ids
 
         # None and empty both mean "no selection" -> never raises.
         _validate_output_ids(None, ["ChatOutput-a"])
         _validate_output_ids([], ["ChatOutput-a"])
 
     def test_all_known_ids_pass(self):
-        from langflow.api.v2.workflow import _validate_output_ids
+        from langflow.api.v2.workflow_validation import _validate_output_ids
 
         _validate_output_ids(["ChatOutput-a"], ["ChatOutput-a", "ChatOutput-b"])
 
     def test_unknown_id_raises_422_listing_available(self):
         from fastapi import HTTPException
-        from langflow.api.v2.workflow import _validate_output_ids
+        from langflow.api.v2.workflow_validation import _validate_output_ids
 
         with pytest.raises(HTTPException) as exc:
             _validate_output_ids(["ChatOutput-z"], ["ChatOutput-a", "ChatOutput-b"])
@@ -813,3 +814,62 @@ class TestValidateOutputIds:
         detail = exc.value.detail
         assert "ChatOutput-z" in str(detail)
         assert detail["available"] == ["ChatOutput-a", "ChatOutput-b"]
+
+
+class TestWorkflowErrorSanitization:
+    """Internal exception text must not leak into client error bodies (I3).
+
+    Each handler logs the full exception server-side and returns a generic,
+    code-tagged message. ``SENTINEL`` stands in for a DSN / driver / stack
+    string that must never reach the client.
+    """
+
+    SENTINEL = "internal-detail-xyzzy-must-not-leak"
+
+    async def test_execute_workflow_db_error_is_generic(self, client: AsyncClient, created_api_key):
+        with patch("langflow.api.v2.workflow.get_flow_by_id_or_endpoint_name") as mock_get_flow:
+            mock_get_flow.side_effect = OperationalError("SELECT 1", {}, Exception(self.SENTINEL))
+            headers = {"x-api-key": created_api_key.api_key}
+            response = await client.post("api/v2/workflows", json={"flow_id": str(uuid4())}, headers=headers)
+        assert response.status_code == 503
+        result = response.json()
+        assert result["detail"]["code"] == "DATABASE_ERROR"
+        assert result["detail"]["message"] == "Failed to fetch flow. Please try again."
+        assert self.SENTINEL not in str(result)
+
+    async def test_execute_workflow_unexpected_error_is_generic(self, client: AsyncClient, created_api_key):
+        with patch("langflow.api.v2.workflow.get_flow_by_id_or_endpoint_name") as mock_get_flow:
+            mock_get_flow.side_effect = ValueError(self.SENTINEL)
+            headers = {"x-api-key": created_api_key.api_key}
+            response = await client.post("api/v2/workflows", json={"flow_id": str(uuid4())}, headers=headers)
+        assert response.status_code == 500
+        result = response.json()
+        assert result["detail"]["code"] == "INTERNAL_SERVER_ERROR"
+        assert result["detail"]["message"] == "An unexpected error occurred."
+        assert self.SENTINEL not in str(result)
+
+    async def test_get_status_db_error_is_generic(self, client: AsyncClient, created_api_key):
+        with patch("langflow.api.v2.workflow.get_job_service") as mock_get_job_service:
+            mock_service = MagicMock()
+            mock_service.get_job_by_job_id = AsyncMock(side_effect=RuntimeError(self.SENTINEL))
+            mock_get_job_service.return_value = mock_service
+            headers = {"x-api-key": created_api_key.api_key}
+            response = await client.get(f"api/v2/workflows?job_id={uuid4()}", headers=headers)
+        assert response.status_code == 500
+        result = response.json()
+        assert result["detail"]["code"] == "INTERNAL_SERVER_ERROR"
+        assert result["detail"]["message"] == "Failed to retrieve job. Please try again."
+        assert self.SENTINEL not in str(result)
+
+    async def test_stop_workflow_db_error_is_generic(self, client: AsyncClient, created_api_key):
+        with patch("langflow.api.v2.workflow.get_job_service") as mock_get_job_service:
+            mock_service = MagicMock()
+            mock_service.get_job_by_job_id = AsyncMock(side_effect=RuntimeError(self.SENTINEL))
+            mock_get_job_service.return_value = mock_service
+            headers = {"x-api-key": created_api_key.api_key}
+            response = await client.post("api/v2/workflows/stop", json={"job_id": str(uuid4())}, headers=headers)
+        assert response.status_code == 500
+        result = response.json()
+        assert result["detail"]["code"] == "INTERNAL_SERVER_ERROR"
+        assert result["detail"]["message"] == "Failed to retrieve job status. Please try again."
+        assert self.SENTINEL not in str(result)
